+++ src/commands/new/index.ts
@@ -161,8 +161,8 @@ export default async function (config: OmniConfig, componentName: string, option
       case 'spa-react':
         newTplPkj = '@omni-door/tpl-spa-react';
         break;
-      case 'component-library-react':
-        newTplPkj = '@omni-door/tpl-component-library-react';
+      case 'component-react':
+        newTplPkj = '@omni-door/tpl-component-react';
         break;
       case 'toolkit':
       default: